allow pyinstaller --onefile to work (at least on win32)

+++ go_timer.py
@@ -9,20 +9,23 @@ from listener.gamestate_listener import ListenerWrapper
 THIS_DIR = os.path.abspath(os.path.dirname(__file__))
 CFG_FILE = os.path.join(THIS_DIR, 'listener', 'gamestate_integration_go_timer.cfg')
 
-def ensure_cfg_file_is_there():
-    # todo... if needed find out more csgo locations
-    possible_paths = ['~/.steam/steamapps/common/Counter-Strike Global Offensive/',
-                     r"C:\Program Files (x86)\Steam\steamapps\common\Counter-Strike Global Offensive"]
-    for p_str in possible_paths:
-        p = pathlib.Path(p_str).expanduser()
-        if p.is_dir():
-            copy_to_path = p / "csgo" / "cfg"
-            if copy_to_path.is_file():
-                copy_to_path.unlink()
+multiprocessing.freeze_support()
 
-            shutil.copy(CFG_FILE, copy_to_path)
-            print ("Copied %s -> %s" % (CFG_FILE, copy_to_path))
-            return True
+def ensure_cfg_file_is_there():
+    if os.path.isfile(CFG_FILE):
+        # todo... if needed find out more csgo locations
+        possible_paths = ['~/.steam/steamapps/common/Counter-Strike Global Offensive/',
+                            r"C:\Program Files (x86)\Steam\steamapps\common\Counter-Strike Global Offensive"]
+        for p_str in possible_paths:
+            p = pathlib.Path(p_str).expanduser()
+            if p.is_dir():
+                copy_to_path = p / "csgo" / "cfg"
+                if copy_to_path.is_file():
+                    copy_to_path.unlink()
+
+                shutil.copy(CFG_FILE, copy_to_path)
+                print ("Copied %s -> %s" % (CFG_FILE, copy_to_path))
+                return True
 
     return False
